+++ inst/htmlwidgets/lib/map/geolocation.js
@@ -1,26 +1,42 @@
 
-function add_geolocation() {
+function add_geolocation(map_id) {
+
+  if(!HTMLWidgets.shinyMode) return;
 
   //var infoWindow = new google.maps.InfoWindow;
   // Try HTML5 geolocation.
+  var eventInfo;
+
   if (navigator.geolocation) {
     navigator.geolocation.getCurrentPosition(function(position) {
 
-      //console.log( position );
-      /*
-      var pos = {
-        lat: position.coords.latitude,
-        lng: position.coords.longitude
-        };
+    console.log( position );
 
-      infoWindow.setPosition(pos);
-      infoWindow.setContent('Location found.');
-      infoWindow.open(map);
-      map.setCenter(pos);
-      */
+    eventInfo = {
+      lat: position.coords.latitude,
+      lon: position.coords.longitude,
+      accuracy: position.coords.accuracy,
+      altitude: position.coords.altitude,
+      altitudeAccuracy: position.coords.altitudeAccuracy,
+      heading: position.coords.heading,
+      speed: position.coords.speed,
+      timestamp: position.timestamp,
+      randomValue: Math.random()
+    };
+
+    console.log( eventInfo );
+
+      //infoWindow.setPosition(pos);
+      //infoWindow.setContent('Location found.');
+      //infoWindow.open(map);
+      //map.setCenter(pos);
+
     event_return_type = window.googleway.params[1].event_return_type;
-    eventInfo = event_return_type === "list" ? position : JSON.stringify(position);
-    Shiny.onInputChange(map_id + "_geolocation", position);
+    eventInfo = event_return_type === "list" ? eventInfo : JSON.stringify(eventInfo);
+
+    console.log( eventInfo );
+
+    Shiny.onInputChange(map_id + "_geolocation", eventInfo);
 
     }, function() {
       //handleLocationError(true, infoWindow, map.getCenter());